Support client-only tasks globally, and per-story.
For projects that don't server render their components, the
server-related data is not helpful.

Clients can opt-in to client-only performance tests by adding:

    import { addParameters } from '@storybook/client-api';

    addParameters({
      performance: {
        clientOnly: true,
      },
    });

to their .storybook/preview.js file.

On a per-story basis, the `clientOnly` parameter can be set alongside
the `interactions` parameter:

    select.story = {
      name: 'React select',
      parameters: {
        performance: {
          interactions: interactionTasks,
          clientOnly: true,
        },
      },
    };

This change removes all service-sid tasks, as well as the "hydrate"
task in from the client-side tasks.